+++ ConcordiumWallet/Views/AccountsView/AccountsCoordinator.swift
@@ -29,9 +29,9 @@ class AccountsCoordinator: Coordinator {
     var navigationController: UINavigationController
     weak var delegate: AccountsCoordinatorDelegate?
     weak var accountsPresenterDelegate: AccountsPresenterDelegate?
-    private var cancellables: Set<AnyCancellable> = []
     private weak var appSettingsDelegate: AppSettingsDelegate?
     private var dependencyProvider: DependencyProvider
+
     init(
         navigationController: UINavigationController,
         dependencyProvider: DependencyProvider,
@@ -42,126 +42,7 @@ class AccountsCoordinator: Coordinator {
         self.dependencyProvider = dependencyProvider
         self.appSettingsDelegate = appSettingsDelegate
         self.accountsPresenterDelegate = accountsPresenterDelegate
-        
-        let metadata = AppMetadata(
-            name: "Concordium",
-            description: "Concordium - Blockchain Wallet",
-            url: "wallet.connect",
-            icons: [],
-            verifyUrl: "verify.walletconnect.com"
-        )
-        Pair.configure(metadata: metadata)
-        Networking.configure(projectId: "76324905a70fe5c388bab46d3e0564dc", socketFactory: SocketFactory())
-        
-        // TODO Define a service for WalletConnect that tracks the currently open sessions (similarly to what dapp-libraries do on the client side...).
-        
-        // TODO Don't allow going back to views corresponding to previous states.
-        
-        // Register handlers for WalletConnect events.
-        
-        // Handler for session proposals, i.e. requests for connections to be established.
-        Sign.instance.sessionProposalPublisher
-            .receive(on: DispatchQueue.main)
-            .sink(receiveCompletion: { failure in
-                print(failure) // TODO: should we handle error?
-            }, receiveValue: { [weak self] proposal, _ in
-                
-                guard let self = self else { return }
-                print("DEBUG: \(self) Session \(proposal) proposed")
 
-                // TODO Auto-reject proposal if namespaces doesn't exactly match expected chain/method/event.
-                //      And show user appropriate error...
-                
-                let viewModel = WalletConnectAccountSelectViewModel(
-                    storageManager: self.dependencyProvider.storageManager(),
-                    proposal: proposal
-                )
-                
-                viewModel.didSelectAccount = { accountAddress in
-                    self.navigationController.pushViewController(
-                        UIHostingController(
-                            rootView: WalletConnectApprovalView(
-                                title: "walletconnect.connect.approve.title".localized,
-                                subtitle: "walletconnect.connect.approve.subtitle".localizedNonempty,
-                                contentView: WalletConnectProposalApprovalView(
-                                    accountName: accountAddress,
-                                    proposal: proposal.proposalData
-                                ),
-                                viewModel: .init(
-                                    didAccept: {
-                                        Task {
-                                            do {
-                                                try await Sign.instance.approve(
-                                                    proposalId: proposal.id,
-                                                    namespaces: [ // TODO un-hardcode
-                                                        "ccd": SessionNamespace(
-                                                            chains: [Blockchain("ccd:testnet")!],
-                                                            accounts: [Account("ccd:testnet:\(accountAddress)")!],
-                                                            methods: ["sign_and_send_transaction", "sign_message"],
-                                                            events: ["chain_changed", "accounts_changed"]
-                                                        )
-                                                                ])
-                                            } catch {
-                                                print("ERROR: approval of connection failed: \(error)")
-                                            }
-                                        }
-                                    },
-                                    didDecline: {
-                                        // User declined the request to connect: Reject it.
-                                        Task {
-                                            do {
-                                                try await Sign.instance.reject(proposalId: proposal.id, reason: .userRejected)
-                                                //                                                try await Sign.instance.disconnect(topic: proposal)
-                                            } catch let error {
-                                                print("ERROR: rejection of connection failed: \(error)")
-                                            }
-                                        }
-                                        
-                                        // Pop the VC without waiting for rejection to complete.
-                                        self.navigationController.setNavigationBarHidden(false, animated: false)
-                                        self.navigationController.popToRootViewController(animated: true)
-                                    }
-                                )
-                            )
-                        ),
-                        animated: true
-                    )
-                }
-                let viewController = WalletConnectAccountSelectViewController(viewModel: viewModel)
-                self.navigationController.setNavigationBarHidden(true, animated: false)
-                self.navigationController.pushViewController(viewController, animated: true)
-                
-            })
-            .store(in: &cancellables)
-        
-        Sign.instance.sessionSettlePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { session in
-                print("DEBUG: Session \(session.pairingTopic) settled")
-                
-                // Connection established: Open "connected" screen.
-                self.navigationController.pushViewController(
-                    UIHostingController(
-                        rootView: WalletConnectConnectedView(
-                            dappName: "TODO", accountName: "TODO", didDisconnect: {
-                                // User clicked the disconnect button.
-                                Task {
-                                    do {
-                                        try await Sign.instance.disconnect(topic: session.topic)
-                                    } catch let error {
-                                        print("ERROR: cannot disconnect: \(error)")
-                                    }
-                                }
-                                // Pop the VC without waiting for disconnect to complete.
-                                self.navigationController.setNavigationBarHidden(false, animated: false)
-                                self.navigationController.popToRootViewController(animated: true)
-                            }
-                        )
-                    ),
-                    animated: true
-                )
-            }
-            .store(in: &cancellables)
     }
 
     func start() {
@@ -371,104 +252,80 @@ protocol WalletConnectDelegate: AnyObject {
 
 extension AccountsCoordinator: WalletConnectDelegate {
     func showWalletConnectScanner() {
-    
         // For now we just print the following events.
-        Sign.instance.sessionDeletePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { sessionId, reason in
-                // Called when the dApp disconnects - not when we do ourselves!
-                print("DEBUG: Session \(sessionId) deleted with reason \(reason)")
-                
-                // Connection lost or disconnected: Pop "connected" screen.
-                // TODO Only do this if we're actually on that screen (i.e. the deleted session matches the currently connected one).
-                self.navigationController.setNavigationBarHidden(false, animated: false)
-                self.navigationController.popToRootViewController(animated: true)
-            }
-            .store(in: &cancellables)
-        Sign.instance.sessionEventPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { event in
-                print("DEBUG: Session event: \(event)")
-            }
-            .store(in: &cancellables)
-        Sign.instance.sessionExtendPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { sessionTopic, date in
-                print("DEBUG: Session \(sessionTopic) extended until \(date)")
-            }
-            .store(in: &cancellables)
-    
-        Sign.instance.sessionUpdatePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { sessionTopic, namespaces in
-                print("DEBUG: Session \(sessionTopic) updated")
-            }
-            .store(in: &cancellables)
-        Sign.instance.socketConnectionStatusPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { status in
-                print("DEBUG: Socket connection status update: \(status)")
-            }
-            .store(in: &cancellables)
-        Sign.instance.sessionResponsePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { response in
-                print("DEBUG: Response: \(response)")
-            }
-            .store(in: &cancellables)
-        Sign.instance.sessionRejectionPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { proposal, reason in
-                print("DEBUG: Proposal \(proposal) rejected with reason \(reason)")
-            }
-            .store(in: &cancellables)
-        
-        Sign.instance.pingResponsePublisher
-            .receive(on: DispatchQueue.main)
-            .sink { ping in
-                print("DEBUG: Ping: \(ping)")
-            }
-            .store(in: &cancellables)
+//        Sign.instance.sessionDeletePublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { sessionId, reason in
+//                // Called when the dApp disconnects - not when we do ourselves!
+//                print("DEBUG: Session \(sessionId) deleted with reason \(reason)")
+//
+//                // Connection lost or disconnected: Pop "connected" screen.
+//                // TODO: Only do this if we're actually on that screen (i.e. the deleted session matches the currently connected one).
+//                self.navigationController.setNavigationBarHidden(false, animated: false)
+//                self.navigationController.popToRootViewController(animated: true)
+//            }
+//            .store(in: &cancellables)
+//
+//        Sign.instance.sessionEventPublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { event in
+//                print("DEBUG: Session event: \(event)")
+//            }
+//            .store(in: &cancellables)
+//        Sign.instance.sessionExtendPublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { sessionTopic, date in
+//                print("DEBUG: Session \(sessionTopic) extended until \(date)")
+//            }
+//            .store(in: &cancellables)
+//
+//        Sign.instance.sessionUpdatePublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { sessionTopic, _ in
+//                print("DEBUG: Session \(sessionTopic) updated")
+//            }
+//            .store(in: &cancellables)
+//        Sign.instance.socketConnectionStatusPublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { status in
+//                print("DEBUG: Socket connection status update: \(status)")
+//            }
+//            .store(in: &cancellables)
+//        Sign.instance.sessionResponsePublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { response in
+//                print("DEBUG: Response: \(response)")
+//            }
+//            .store(in: &cancellables)
+//        Sign.instance.sessionRejectionPublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { proposal, reason in
+//                print("DEBUG: Proposal \(proposal) rejected with reason \(reason)")
+//            }
+//            .store(in: &cancellables)
+//
+//        Sign.instance.pingResponsePublisher
+//            .receive(on: DispatchQueue.main)
+//            .sink { ping in
+//                print("DEBUG: Ping: \(ping)")
+//            }
+//            .store(in: &cancellables)
+//
+//        setupWalletConnectRequestBinding()
+//
+////        // Temporarily use hardcoded connection string rather than scanning QR code.
+////        // Unsure why, but if we clear pairings and instantiate this one, it seems to connect without the proposal thing...
+////        let wc = "wc:39e17f8223d80748d56e528c6715f211afd3b4e0dee4ac887a79d644308d157a@2?relay-protocol=irn&symKey=c370924b1dd87b047fc8726f33d80adecc41f350635979bfbe597c58fcf5f2bd"
+////        do {
+////            try Pair.instance.cleanup()
+////        } catch let error {
+////            print("ERROR: cannot clean up pairings: \(error)")
+////        }
 
-        // Handler for incoming requests on established connection.
-        Sign.instance.sessionRequestPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { request, _ in
-                print("DEBUG: Incoming request: \(request)") // TODO handle...
-            }
-            .store(in: &cancellables)
-
-//        // Temporarily use hardcoded connection string rather than scanning QR code.
-//        // Unsure why, but if we clear pairings and instantiate this one, it seems to connect without the proposal thing...
-//        let wc = "wc:39e17f8223d80748d56e528c6715f211afd3b4e0dee4ac887a79d644308d157a@2?relay-protocol=irn&symKey=c370924b1dd87b047fc8726f33d80adecc41f350635979bfbe597c58fcf5f2bd"
-//        do {
-//            try Pair.instance.cleanup()
-//        } catch let error {
-//            print("ERROR: cannot clean up pairings: \(error)")
-//        }
-        
-        // Show QR code scanner.
-        let vc = ScanQRViewControllerFactory.create(
-            with: ScanQRPresenter(
-                didScanQrCode: { [weak self] value in
-                    // TODO Can do more detailed check?
-                    if !value.hasPrefix("wc:") {
-                        return false
-                    }
-
-                    // Successfully scanner WalletConnect QR.
-                    Task {
-                        do {
-                            try await Pair.instance.pair(uri: WalletConnectURI(string: value)!)
-                        } catch let error {
-                            print("ERROR: cannot pair: \(error)")
-                        }
-                    }
-                    self?.navigationController.popViewController(animated: true)
-                    return true
-                }
-            )
+        let coordinator = WalletConnectCoordinator(navigationController: navigationController,
+                                                   dependencyProvider: dependencyProvider,
+                                                   parentCoodinator: self
         )
-        navigationController.pushViewController(vc, animated: true)
+        coordinator.start()
     }
 }
