+++ ConcordiumWallet/Views/WalletConnect/AccountSelect/WalletConnectAccountSelectView.swift
@@ -14,9 +14,9 @@ struct WalletConnectAccountSelectView: View {
         Text("walletconnect.select.account.header".localized)
             .multilineTextAlignment(.center)
             .padding(32)
-        List($viewModel.accounts, id: \.address) { account in
+        List($viewModel.accounts, id: \.accountIndex) { account in
             WalletConnectAccountItemView(account: account.wrappedValue) {
-                viewModel.didSelect(accountAddress: account.wrappedValue.address)
+                viewModel.didSelect(account: account.wrappedValue)
             }
         }
         .listRowSeparator(.hidden)
